UI - PieException Exception class is moved into the exceptions subfolder.
