+++ src/main/java/sdmExam/exceptions/PieException.java
@@ -1,4 +1,4 @@
-package sdmExam.UI.Exception;
+package sdmExam.exceptions;
 
 public class PieException extends  Exception{
     public PieException(String error){
